Resolving connection leak in VirDB implementation
The root cause of the connection leak was in the VirDB constructor:
after the connection was acquired, the constructor was attempting to
perform a query on the connection straight away, however under certain
circumstances, that could result in an exception preventing the object
construction. Since the object was never fully constructed, the
try-with-resources statement had no chance of actually calling the close
method to free up the acquired JDBC connection.
I've modified the initialization logic, so now the initial query is done
separately, also did some cleanup around the debug messages as they weren't
quite descriptive.